requirements.txt
Django==4.2
django-environ==0.9.0 
psycopg2==2.9.4
sqlparse==0.4.3
ansible-runner
djangorestframework-simplejwt
python-ldap
python3-ldap
django-auth-ldap
psycopg2-binary
celery
pytz
asgiref
backports.zoneinfo==0.2.1